+++ src/Paket.Core/Utils.fs
@@ -308,6 +308,7 @@ let inline createWebClient (url,auth:Auth option) =
 
 open System.Diagnostics
 open System.Threading
+open System.Collections.Generic
 
 let innerText (exn:Exception) =
     match exn.InnerException with
@@ -498,21 +499,25 @@ let inline (++) x y =
 
 let parseKeyValuePairs (s:string) =
     let s = s.Trim()
-    let parts = s.Split([|','|], StringSplitOptions.RemoveEmptyEntries)
-    let dict = new System.Collections.Generic.Dictionary<_,_>()
+    try
+        let parts = s.Split([|','|], StringSplitOptions.RemoveEmptyEntries)
+        let dict = Dictionary<_,_>()
 
-    let lastKey = ref ""
+        let lastKey = ref ""
 
-    for p in parts do
-        if p.Contains ":" then
-            let innerParts = p.Split ':' |> Array.map String.trim
-            if innerParts.Length % 2 <> 0 then
-                failwithf "\"%s\" can not be parsed as key/value pairs." p
-            dict.Add(innerParts.[0].ToLower(),innerParts.[1])
-            lastKey := innerParts.[0]
-        else
-            dict.[!lastKey] <- dict.[!lastKey] + ", " + p
-    dict
+        for p in parts do
+            if p.Contains ":" then
+                let innerParts = p.Split ':' |> Array.map String.trim
+                if innerParts.Length % 2 <> 0 then
+                    failwithf "\"%s\" can not be parsed as key/value pairs." p
+                dict.Add(innerParts.[0].ToLower(),innerParts.[1])
+                lastKey := innerParts.[0]
+            else
+                dict.[!lastKey] <- dict.[!lastKey] + ", " + p
+        dict
+    with
+    | exn -> 
+        failwithf "Could not parse %s as key/value pairs.%s%s" s Environment.NewLine exn.Message
 
 let downloadStringSync (url : string) (client : System.Net.WebClient) = 
     try